Simple calculator fixed

+++ src/main/java/lv/rvt/App.java
@@ -18,7 +18,7 @@ public class App
         System.out.println(num1 + " - " + num2 + " = " + result2);
         double result3 = num1 * num2;
         System.out.println(num1 + " * " + num2 + " = " + result3);
-        double result4 = num1 / num2;
+        double result4 = (double)num1 / num2;
         System.out.println(num1 + " / " + num2 + " = " + result4);
 
     }
